% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bed_to_bedpe}
\alias{bed_to_bedpe}
\title{Convert bed format to bedpe format}
\usage{
bed_to_bedpe(bed)
}
\arguments{
\item{bed}{SV data in bed format}
}
\value{
SV data in bedpe format
}
\description{
This function read SV data in bed format and convert it into bedpe format
}
