% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Check_bedtools}
\alias{Check_bedtools}
\title{Check the directory of bedtools}
\usage{
Check_bedtools(x = "bedtools")
}
\arguments{
\item{x}{"bedtools"}
}
\value{
bedtools path
}
\description{
This function check the binary of bedtools
}
