% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairtopair_filter}
\alias{pairtopair_filter}
\title{Filter SV types and breakpoint positions}
\usage{
pairtopair_filter(pairtopair, PASS_filter, svtype_ignore)
}
\arguments{
\item{pairtopair}{file name}

\item{PASS_filter}{filtering of FILTER field in VCF of two calls}

\item{svtype_ignore}{whether ignore SV type for integration}
}
\value{
data frame
}
\description{
This function read bed format
}
