% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{plot_ideograms}
\alias{plot_ideograms}
\title{Plot genomic bins hotspots}
\usage{
plot_ideograms(df_bin_all_hotspots)
}
\arguments{
\item{df_bin_all_hotspots}{data frame of SV breakpoints bins with hotspots defined}
}
\value{
data frame of SV set
}
\description{
Plot hotspots
}
