% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Prepare_for_VennDiagram}
\alias{Prepare_for_VennDiagram}
\title{Prepare_for_VennDiagram}
\usage{
Prepare_for_VennDiagram(SV_integrated, sampleID, SVCaller_name)
}
\arguments{
\item{SV_integrated}{integrated SV call set}

\item{sampleID}{sample ID}

\item{SVCaller_name}{name of callers}
}
\value{
data frame
}
\description{
This function read bed format
}
