% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_SV_for_circos}
\alias{prepare_SV_for_circos}
\title{Prepare data for SV circos plot}
\usage{
prepare_SV_for_circos(bedpe)
}
\arguments{
\item{bedpe}{SV data in bedpe format, for example output from simple_SVTYPE_classification}
}
\value{
list of input data for circlize package
}
\description{
This function read SV data in bedpe format and convert to list of input data for circlize package
}
