% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{spectrum_SV_bin}
\alias{spectrum_SV_bin}
\title{Genomic bins}
\usage{
spectrum_SV_bin(df_breakpoints)
}
\arguments{
\item{df_breakpoints}{data frame of SV breakpoints}
}
\value{
data frame of SV set
}
\description{
Put SV breakpoint into bins
}
