% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{Spectrum_SV_bin_generate}
\alias{Spectrum_SV_bin_generate}
\title{Genomic bins hotspots}
\usage{
Spectrum_SV_bin_generate(
  All_sampleID,
  All_input_df_name,
  threshold_count_breakpoint,
  threshold_count_sample
)
}
\arguments{
\item{All_sampleID}{sample ID for all samples}

\item{All_input_df_name}{names of all bed in df}

\item{threshold_count_breakpoint}{threshold of number of SD}

\item{threshold_count_sample}{threshold of number of samples}
}
\value{
data frame of genomic bins with hotspots defined
}
\description{
Define hotspots
}
