% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{spectrum_SV_type}
\alias{spectrum_SV_type}
\title{Count of SV types in cohort}
\usage{
spectrum_SV_type(
  All_sampleID,
  All_SV_data,
  identify_hyperSV_tumour = FALSE,
  threshold_total = NULL,
  threshold_relative_freq = NULL
)
}
\arguments{
\item{All_sampleID}{sample ID}

\item{All_SV_data}{list of SV files or data frames}

\item{threshold_total}{threshold of minimum total count of SVs per sample}

\item{threshold_relative_freq}{threshold of minimum relative frequency of one SV type}

\item{identify_hyperSV_tumour:}{TRUE or FALSE. Whether to identify hyper-SV mutated tumour samples for large cancer cohort. Default as FALSE.}
}
\value{
data frame of hyper SV
}
\description{
This function count SV types for cohort and optionally identify the hyper-SV subgroup
}
