% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{spectrum_SV_type_plot}
\alias{spectrum_SV_type_plot}
\title{Spectrum of SV figures}
\usage{
spectrum_SV_type_plot(input_SV_count)
}
\arguments{
\item{input_SV_count}{sample ID}
}
\value{
figures
}
\description{
This function read bed format
}
