% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standard_bedtool_prepare_bkpt}
\alias{standard_bedtool_prepare_bkpt}
\title{Prepare breakpoint to use for bedtools}
\usage{
standard_bedtool_prepare_bkpt(SV_data, bkpt_T_callers, caller)
}
\arguments{
\item{SV_data}{data frame}

\item{caller}{caller name or unique identifier of VCF}

\item{bkpt_T_caller}{the maximun value of breakpoint difference}
}
\value{
data frame
}
\description{
This function read bed format
}
