% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strands_standardisation}
\alias{strands_standardisation}
\title{Convert strands into standard form}
\usage{
strands_standardisation(bed)
}
\arguments{
\item{bed}{dataframe}
}
\value{
data frame
}
\description{
This function read bed format
}
