% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{Summary_gene_fusions}
\alias{Summary_gene_fusions}
\title{Define frequency of gene fusions}
\usage{
Summary_gene_fusions(bed_GeneAnnotated)
}
\arguments{
\item{bed_GeneAnnotated}{data frame}
}
\value{
data frame of summarised gene fusions
}
\description{
Define frequency of gene fusions
}
