% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{summary_SV_breakpoint}
\alias{summary_SV_breakpoint}
\title{SV breakpoints summary for all samples}
\usage{
summary_SV_breakpoint(All_sampleID, All_SV_data)
}
\arguments{
\item{All_sampleID}{sample ID for all samples}

\item{All_SV_data}{vcf files or a list of dataframe of SV data}
}
\value{
data frame of SV set
}
\description{
This function put all sample SV breakpoints together
}
