% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{summary_SV_type}
\alias{summary_SV_type}
\title{Count of SV types in cohort}
\usage{
summary_SV_type(All_sampleID, All_SV_data)
}
\arguments{
\item{All_sampleID}{sample ID}

\item{SVdf_list}{list of SV data.frames}
}
\value{
data frame of count of SVs in cohort
}
\description{
This function count SV types for more than one sample in a cohort
}
