% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{SV_bedpe_gene_annotation}
\alias{SV_bedpe_gene_annotation}
\title{SV breakpoints gene annotation on input bed format}
\usage{
SV_bedpe_gene_annotation(input_df_name, gene_bed, bedtools_dir)
}
\arguments{
\item{input_df_name}{name of input data frame}

\item{gene_bed}{data frame of gene regions}

\item{bedtools_dir}{bedtools for use}
}
\value{
data frame of SV set
}
\description{
This function annotate SV breakpoints based on gene regions
}
