% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{SV_gene_annotation}
\alias{SV_gene_annotation}
\title{SV breakpoints gene annotation on input bed format}
\usage{
SV_gene_annotation(SV_data, gene_bed, bedtools_dir)
}
\arguments{
\item{SV_data}{input SV data frame}

\item{gene_bed}{gene regions in data.frame}

\item{bedtools_dir}{bedtools for use}
}
\value{
data frame of SV set
}
\description{
This function annotate SV breakpoints based on gene regions
}
