% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_integration_filtration.R
\name{SV_integration}
\alias{SV_integration}
\title{Integrate SV call sets}
\usage{
SV_integration(
  vcf_files,
  SVCaller_names,
  sampleID = "sample_1",
  bkpt_T_callers = 100,
  PASS_filter = "both",
  svtype_ignore = FALSE,
  bedtools_dir = NULL
)
}
\arguments{
\item{vcf_files}{list of VCF files from different callers}

\item{SVCaller_names}{names of callers or unique identifier of SV set}

\item{sampleID}{unique identifier of sample, default as "sample_1".}

\item{bkpt_T_callers}{threshold of breakpoint difference. Default as 100 bp}

\item{PASS_filter}{filtering based on FILTER field of two calls in VCF: "both" to require both calls with "PASS", one to require one of the two calls with "PASS", "none" to ignore this filtering. Default is "both".}

\item{svtype_ignore}{whether ignore SV type for integration. Default as FALSE.}

\item{bedtools_dir}{directory of bedtools}
}
\value{
SV data in bedpe format
}
\description{
This function read SV data and output the union set of them.
}
