% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_integration_filtration.R
\name{SV_integration_pairwise}
\alias{SV_integration_pairwise}
\title{Integrate SV call sets and write output}
\usage{
SV_integration_pairwise(
  sampleID,
  df_SV,
  vcf_files,
  bkpt_T_callers,
  SVCaller_names,
  PASS_filter,
  svtype_ignore,
  bedtools_dir
)
}
\arguments{
\item{sampleID}{unique identifier of sample}

\item{df_SV}{name of callers}

\item{vcf_files}{list of VCF files from different callers}

\item{bkpt_T_callers}{threshold of breakpoint difference}

\item{SVCaller_names}{names of callers}

\item{PASS_filter}{filtering based on FILTER field of two calls in VCF: "both" to require both calls with "PASS", one to require one of the two calls with "PASS", "none" to ignore this filtering. Default is "both".}

\item{svtype_ignore}{whether ignore SV type for integration}

\item{bedtools_dir}{directory of bedtools}
}
\value{
data frame
}
\description{
This function read bed format
}
