% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SVCaller_bed_newID_generate}
\alias{SVCaller_bed_newID_generate}
\title{Integrate SV call sets}
\usage{
SVCaller_bed_newID_generate(bedpe, SVCaller_name = "SVUtil")
}
\arguments{
\item{bedpe}{SV set}

\item{SVCaller_name}{name of callers}
}
\value{
data frame
}
\description{
This function read bed format
}
