% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SVCaller_union_intersect_generate}
\alias{SVCaller_union_intersect_generate}
\title{Integrate SV call sets}
\usage{
SVCaller_union_intersect_generate(
  sampleID,
  df_SV,
  vcf_files,
  bkpt_T_callers,
  SVCaller_names,
  bedtools_dir,
  DIR
)
}
\arguments{
\item{sampleID}{sample ID}

\item{df_SV}{name of callers}

\item{vcf_files}{names of call sets}

\item{bkpt_T_callers}{threshold of breakpoint difference}

\item{SVCaller_names}{whether ignore SV type for integration}

\item{bedtools_dir}{directory of bedtools}

\item{DIR}{directory to write temporary files in}
}
\value{
data frame
}
\description{
This function read bed format
}
