% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{SVTYPE_stat_generate}
\alias{SVTYPE_stat_generate}
\title{Count the SV types}
\usage{
SVTYPE_stat_generate(bedpe)
}
\arguments{
\item{bedpe}{SV data in bedpe format}
}
\value{
data frame of SV count in each type category
}
\description{
This function read SV data in bedpe format and count SV types
}
