% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TypePosfilter}
\alias{TypePosfilter}
\title{Check SV types and breakpoint positions}
\usage{
TypePosfilter(intersect_file, SVTYPE_ignore)
}
\arguments{
\item{intersect_file}{file name}

\item{SVTYPE_ignore}{whether ignore SV type for integration}
}
\value{
data frame
}
\description{
This function read bed format
}
