% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bed_to_bedpe}
\alias{bed_to_bedpe}
\title{Convert bed format to bedpe format}
\usage{
bed_to_bedpe(bed)
}
\arguments{
\item{bed}{data frame}
}
\value{
data frame
}
\description{
This function read bed format
}
