% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinySoSV_prediction.R
\name{ShinySoSV_prediction}
\alias{ShinySoSV_prediction}
\title{ShinySoSV prediction}
\usage{
ShinySoSV_prediction(
  Candidate_callers,
  newdata,
  performance,
  callset,
  ShinySoSV_DIR
)
}
\arguments{
\item{newdata}{data frame of variables with which to predict}

\item{performance}{should be any of "sensitivity", "precision" or "F1 score"}

\item{callset}{individual caller, pairwise union or intersection}

\item{ShinySoSV_DIR}{directory of installed ShinySoSV}

\item{candidate_callers}{names of callers}
}
\value{
data frame of predicted performance
}
\description{
This function predict the SV calling performance of SV callers and their combinations
}
