% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_integration_filtration.R
\name{simple_SVTYPE_classification}
\alias{simple_SVTYPE_classification}
\title{Classify SV types based on VCF}
\usage{
simple_SVTYPE_classification(vcf_file, caller_name)
}
\arguments{
\item{vcf_file}{names of vcf file}

\item{caller_name}{name of caller to be used in ID}
}
\value{
data frame
}
\description{
This function read bed format
}
