% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_integration_filtration.R
\name{simple_SVTYPE_classification}
\alias{simple_SVTYPE_classification}
\title{Classify SV types based on VCF}
\usage{
simple_SVTYPE_classification(SV_data, caller_name = "caller1")
}
\arguments{
\item{SV_data}{SV VCF path or SV in data.frame}

\item{caller_name}{name of caller to be used in ID}
}
\value{
data frame
}
\description{
This function read bed format
}
