% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{spectrum_SV_bin_define_hotspot}
\alias{spectrum_SV_bin_define_hotspot}
\title{Genomic bins hotspots}
\usage{
spectrum_SV_bin_define_hotspot(
  df_breakpoints,
  threshold_count_breakpoint = NULL,
  threshold_count_sample = NULL
)
}
\arguments{
\item{df_breakpoints}{data frame of SV breakpoints}

\item{threshold_count_breakpoint}{threshold of number of SD}

\item{threshold_count_sample}{threshold of number of samples}
}
\value{
data frame of SV set
}
\description{
Define hotspots
}
