% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{Spectrum_SV_type}
\alias{Spectrum_SV_type}
\title{Spectrum of SV count}
\usage{
Spectrum_SV_type(
  All_sampleID,
  All_input_df_name,
  threshold_total,
  threshold_relative_freq
)
}
\arguments{
\item{All_sampleID}{sample ID}

\item{All_input_df_name}{names}

\item{threshold_total}{threshold of minimum total count of SVs per sample}

\item{threshold_relative_freq}{threshold of minimum relative frequency of one SV type}
}
\value{
data frame of hyper SV
}
\description{
This function read bed format
}
