% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{spectrum_SV_type_plot}
\alias{spectrum_SV_type_plot}
\title{Plot the figures of SV type spectrum}
\usage{
spectrum_SV_type_plot(input_SV_count)
}
\arguments{
\item{input_SV_count}{data.frame of SV count}
}
\value{
list of figures
}
\description{
This function read data.frame of SV count (e.g. from function spectrum_SV_type)
}
