% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Standard_bedtool_prepare_bkpt}
\alias{Standard_bedtool_prepare_bkpt}
\title{Prepare breakpoint to use for bedtools}
\usage{
Standard_bedtool_prepare_bkpt(standard_bedpe, BND_diff)
}
\arguments{
\item{standard_bedpe}{data frame}

\item{BND_diff}{the maximun value of breakpoint difference}
}
\value{
data frame
}
\description{
This function read bed format
}
