% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strands_standardisation}
\alias{strands_standardisation}
\title{Convert strands in VCF into BED}
\usage{
strands_standardisation(bed)
}
\arguments{
\item{bed}{SV in bed format}
}
\value{
SV in bed format
}
\description{
This function read SV in bed format and standardise the strands
}
