% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{Summary_SV_breakpoint}
\alias{Summary_SV_breakpoint}
\title{SV breakpoints summary for all samples}
\usage{
Summary_SV_breakpoint(All_sampleID, All_input_df_name)
}
\arguments{
\item{All_sampleID}{sample ID for all samples}

\item{All_input_df_name}{names of all bed in df}
}
\value{
data frame of SV set
}
\description{
This function put all sample SV breakpoints together
}
