% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{summary_SV_type}
\alias{summary_SV_type}
\title{Summary SV count}
\usage{
summary_SV_type(All_sampleID, SVdf_list)
}
\arguments{
\item{All_sampleID}{sample ID}

\item{SVdf_list}{list of SV data.frames}
}
\value{
data frame
}
\description{
This function read bed format
}
