% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{Summary_SV_type}
\alias{Summary_SV_type}
\title{Summary SV count}
\usage{
Summary_SV_type(All_sampleID, vcf_list)
}
\arguments{
\item{All_sampleID}{sample ID}

\item{vcf_list}{names}
}
\value{
data frame
}
\description{
This function read bed format
}
