% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{SV_breakpoint_gene_annotation}
\alias{SV_breakpoint_gene_annotation}
\title{SV breakpoints gene annotation}
\usage{
SV_breakpoint_gene_annotation(SV_data, gene_bed, bedtools_dir)
}
\arguments{
\item{gene_bed}{data frame of gene regions}

\item{bedtools_dir}{directory of bedtools to use}

\item{SV_date}{input SV data frame}
}
\value{
data frame of intersection set between SV and gene regions
}
\description{
This function annotate SV breakpoints based on gene regions
}
