% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{SV_CNV_integration}
\alias{SV_CNV_integration}
\title{Integrate SV and CNV}
\usage{
SV_CNV_integration(sampleID, SV_bed, CNV_bed, overlap_f, bedtools_dir = NULL)
}
\arguments{
\item{sampleID}{name of sample}

\item{SV_bed}{data frame of SV}

\item{CNV_bed}{data frame of CNV segment}

\item{overlap_f}{the fraction of minimum overlap required of CNV segment as a fraction of SV}

\item{bedtools_dir}{bedtools for use}
}
\value{
data frame of SV set
}
\description{
This function integrate deletion and duplication with copy number segments
}
