% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SVCaller_union_intersect_generate}
\alias{SVCaller_union_intersect_generate}
\title{Integrate SV call sets}
\usage{
SVCaller_union_intersect_generate(
  sampleID,
  SVCaller_name,
  SVCaller_bed_name,
  BND_diff,
  bkpt_T_callers,
  SVTYPE_ignore,
  bedtools_dir,
  sub_directory
)
}
\arguments{
\item{sampleID}{sample ID}

\item{SVCaller_name}{name of callers}

\item{SVCaller_bed_name}{names of call sets}

\item{BND_diff}{maximun breakpoint difference}

\item{bkpt_T_callers}{threshold of breakpoint difference}

\item{SVTYPE_ignore}{whether ignore SV type for integration}

\item{bedtools_dir}{directory of bedtools}

\item{sub_directory}{directory to write temporary files in}
}
\value{
data frame
}
\description{
This function read bed format
}
