% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_tertiary_analysis.R
\name{SVTYPE_stat_generate}
\alias{SVTYPE_stat_generate}
\title{SVtype count}
\usage{
SVTYPE_stat_generate(bedpe)
}
\arguments{
\item{bedpe}{bed format}
}
\value{
data frame
}
\description{
This function read bed format
}
