% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_integration_filtration.R
\name{vcf_to_dataframe}
\alias{vcf_to_dataframe}
\title{Convert VCF format to bed format}
\usage{
vcf_to_dataframe(vcf_file)
}
\arguments{
\item{vcf_file}{names of vcf file}
}
\value{
data frame
}
\description{
This function read vcf file and convert it to bed format
}
